fix contact page display - some quality fixes
