+++ view/frontend/web/css/source/_module.less
@@ -756,12 +756,12 @@
 
                 .image-wrapper-header {
                     width: 100%;
+                    height: 200px;
+                    overflow: hidden;
 
                     img {
-                        height: auto;
                         max-width: 100%;
                         width: 100%;
-                        max-height: 200px;
                     }
                 }
             }
@@ -894,11 +894,37 @@ a.geolocalize:before {
     overflow: hidden;
     text-align: center;
 }
+
+//
+//  Shop Contact Form
+//  _____________________________________________
+
+.form-contact-wrapper {
+    display: flex;
+    flex-wrap: wrap;
+
+    > .shop-details-wrapper,
+    > .form.contact {
+        width: 48%;
+        margin: 0 2% 0 0;
+    }
+}
+
+
 //
 //  Mobile
 //  _____________________________________________
 
 .media-width(@extremum, @break) when (@extremum = 'max') and (@break = @screen__s) {
+
+    .form-contact-wrapper {
+        > .shop-details-wrapper,
+        > .form.contact {
+            width: 90%;
+            margin: 0 0 0 5%;
+        }
+    }
+
     #top-storelocator-chooser {
         margin-top: @indent__s;
         margin-right: 5px;
@@ -1363,12 +1389,12 @@ a.geolocalize:before {
 
         .image-wrapper-header {
             width: 100%;
+            height: 200px;
+            overflow: hidden;
 
             img {
-                height: auto;
                 max-width: 100%;
                 width: 100%;
-                max-height: 200px;
             }
         }
