+++ view/frontend/web/css/source/store-view/_store-view.less
@@ -80,12 +80,12 @@
 
   .image-wrapper-header {
     width: 100%;
+    height: 200px;
+    overflow: hidden;
 
     img {
-      height: auto;
       max-width: 100%;
       width: 100%;
-      max-height: 200px;
     }
   }
